Back-end/pyproject.toml
[tool.poetry]
name = "nexus-education-backend"
version = "0.1.0"
description = "Backend API para Nexus Education - Sistema de análise de ementas"
authors = ["Nexus Education"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
supabase = "^2.0.3"
groq = "^0.4.0"
pypdf2 = "^3.0.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

